requirements.txt
PyMySQL==0.7.6
flask-login==0.4.0
Flask-Bootstrap
flask==0.11.1
Flask-Mail==0.9.1
Flask-Migrate==1.1.0
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
Jinja2==2.8
WTForms==2.1
Flask-PageDown
MarkDown
Bleach
ForgeryPy==0.1